Enhanced pydtf capabilities, Python modules, and new "apktool"
As dtf moves away from the shell, its python capabilities will
slowly increase.  This commit adds "the dtfadb" and "dtfcore"
components of pydtf, which lets you (right now) launch other
modules from python, and use adb commands from a python script.
In addition, python modules can now be installed (more on that
in future commits).  "apktool" was updated to a custom build
that should *work* on API 20W and L, but not perfectly.