Add and use new dtf.globals functionality The "globals" class has been expanded to support a property file "~/.dtf/included/globals.ini" that contains key/value pairs. The goal of this API is to support overall dtf configuration. Right now, the API only supports getting properties.
This initial release adds 2 sections: Client and Bindings

The "Client" section is used to configure details about the
dtfClient. The "Bindings" section is used to configure the location
of bundled tools such as apktool, aapt, baksmali, etc. This allows
you to override the default locations of these files. More
sections to follow!