Add new decorator for sub command calling + launch_dir for module
This is the initial commit of the new @sub_cmd() decorator for calling
sub functions directly. This is part 1/2. The second part will add
a built in usage().

This also adds the cd_launch_dir() method to the Module class, for
CDing to the directory from which the module was launched (not TOP)